refactor(utils): put a function useed in multiple classes in the utils part
